+++ src/help/BUILD
@@ -10,7 +10,6 @@ go_library(
         "//src/cli",
         "//src/core",
         "//src/parse/asp",
-        "//src/utils",
         "//third_party/go:deferred_regex",
         "//third_party/go:go-flags",
         "//third_party/go:logging",
